+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+### 1.2.0
+
+- Add coverImageToCanvas option, allowing overflow (at least one side fits)
+
 ### 1.1.2
 
-- In fitImageToCanvas, change the image placement from centered to top-left corner.
+- In fitImageToCanvas, change the image placement from centered to top-left corner
 
 ### 1.1.1